requirements.txt
# ---------------------------------------------------------------------------
# UK Job Market Intelligence Engine — Python dependencies
# ---------------------------------------------------------------------------

# Web framework & ASGI server
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
httpx>=0.27.0          # FastAPI TestClient + async HTTP

# Dashboard
streamlit>=1.33.0
plotly>=5.21.0

# Database
sqlalchemy>=2.0.29
psycopg2-binary         # version chosen automatically for your Python version

# HTTP client (scrapers)
requests>=2.31.0
urllib3>=2.0.0

# Environment / config
python-dotenv>=1.0.1

# NLP
spacy>=3.7.0

# Data processing
pandas>=2.0.0
numpy>=1.26.0

# Data validation
pydantic>=2.7.0

# Testing
pytest>=8.0.0
pytest-asyncio>=0.23.0

# Retry / resilience (used by urllib3 adapter)
certifi>=2024.2.2
